changed rating from double to float wtf

+++ src/main/java/net/aktivreisen24/model/Vacation.java
@@ -7,14 +7,14 @@ public class Vacation {
     private Short houseNumber;
     private String location;
 
-    private double rating;
+    private float rating;
     private String generelInfo;
     private String description;
 
     //availabilty und comments fehlt
 
 
-    public Vacation(long id, String street, Short houseNumber, String location, double rating, String generelInfo, String description) {
+    public Vacation(long id, String street, Short houseNumber, String location, float rating, String generelInfo, String description) {
         this.id = id;
         this.street = street;
         this.houseNumber = houseNumber;
@@ -37,7 +37,7 @@ public class Vacation {
         return location;
     }
 
-    public double getRating() {
+    public float getRating() {
         return rating;
     }
 
@@ -61,7 +61,7 @@ public class Vacation {
         this.location = location;
     }
 
-    public void setRating(double rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
